(split)http://bugzilla.zimbra.com/show_bug.cgi?id=8181 bug: 8181 (checkpoint)
Copied from Perforce
 Change: 27636

+++ src/java/com/zimbra/cs/zclient/ZTag.java
@@ -42,6 +42,6 @@ public interface ZTag {
      */
     public int getUnreadCount();
 
-    public byte getColor();
+    public int getColor();
 
 }
